Remove empty mv_arg from more-list links
The more_link_template routine calls tag_area with a
hard-coded empty string ('') as the second parameter.
This is used in tag_area as $arg and if defined, is
appended to the generated URL as mv_arg.  The result
is that all more-links are generated with an
empty 'mv_arg='.  This patch replaces the hard-coded
empty string with 'undef' so that the more-links
will no longer carry the empty mv_arg.

Developers desiring to specify an mv_arg can do so,
for example, by adding something along the lines
of form=|mv_arg=whatever| to their [query] or
[search-region] tags.